+++ examples/src/bin/static_basalt_app.rs
@@ -1,11 +1,9 @@
 extern crate basalt;
 extern crate once_cell;
 
-use basalt::{
-	input::MouseButton,
-	interface::bin::{self, BinPosition, BinStyle},
-	Basalt,
-};
+use basalt::input::MouseButton;
+use basalt::interface::bin::{self, BinPosition, BinStyle};
+use basalt::Basalt;
 use once_cell::sync::OnceCell;
 use std::sync::Arc;